sdcardfs: fix space leak
Don't keep looked up dentries around because there's no notification
when the lowerfs unlinks an inode, to collect the upper dentry cache so
the lower inode reference can be released and the storage space
released.

This effectively disables the dcache lookups but it should still be much
faster than fuse sdcardfs.

The alternative is to add a notification mechanism to keep two separate
dcache layers in sync which isn't trivial, or stop ever touching the
lower fs and remove that path.replace from VolumeInfo.java.

Change-Id: I211bd676834126f6f65b3d09ebe951d0375d7985
Signed-off-by: Kevin F. Haggerty <haggertk@lineageos.org>
Signed-off-by: Ryan Andri <ryanandri@linuxmail.org>